/ ‘forkbomb/draw.hs’ / ‘forkbomb/test.png’

+++ forkbomb/draw.hs
@@ -23,7 +23,7 @@ main = do handle <- openFile "10.txt" ReadMode
               fill
               setSourceRGB 0 0 0
               let rects = filter (\(_, _, _, v) -> v == '1') $ rleXY cols contents
-              mapM_ (\(x,y,l,_) -> rectangle (cellw* (fromIntegral $ x)) (cellh*y) (cellh) (cellw*l)) rects
+              mapM_ (\(x,y,l,_) -> rectangle (cellw* (fromIntegral $ x)) (cellh* (fromIntegral y)) (cellh) (cellw*(fromIntegral l))) rects
               return ()
             surfaceWriteToPNG surf "test.png"
